// apps/integration_service/src/multicall/multicall.enum.ts
export enum MulticallContractFunctionEnum {
  lpToken = 'lpToken',
  poolInfo = 'poolInfo',
  poolLength = 'poolLength',
  userInfo = 'userInfo',
  pendingSushi = 'pendingSushi',
  SUSHI = 'SUSHI',
}

export enum MulticallTokenFunctionEnum {
  earned = 'earned',
  totalSupply = 'totalSupply',
  token0 = 'token0',
  token1 = 'token1',
  userInfo = 'userInfo',
  getReserves = 'getReserves',
  decimals = 'decimals',
  name = 'name',
  symbol = 'symbol',
}